Let unwrap2 work with double instead of float

+++ BrillouinAcquisition/external/unwrap/unwrap2D.c
@@ -7,7 +7,7 @@
 //published in the Journal Applied Optics, Vol. 41, No. 35, pp. 7437, 2002.
 //This program was written by Munther Gdeisat, Liverpool John Moores University, United Kingdom.
 //Date 26th August 2007
-//The wrapped phase map is assumed to be of floating point data type. The resultant unwrapped phase map is also of floating point type.
+//The wrapped phase map is assumed to be of double data type. The resultant unwrapped phase map is also of double type.
 //The mask is of byte data type.
 //When the mask is 1 this means that the pixel is valid
 //When the mask is 0 this means that the pixel is invalid (noisy or corrupted pixel)
@@ -15,7 +15,7 @@
 
 #include "unwrap2D.h"
 
-yes_no find_pivot(EDGE *left, EDGE *right, float *pivot_ptr)
+yes_no find_pivot(EDGE *left, EDGE *right, double *pivot_ptr)
 {
   EDGE a, b, c, *p;
 
@@ -49,7 +49,7 @@ yes_no find_pivot(EDGE *left, EDGE *right, float *pivot_ptr)
   return no;
 }
 
-EDGE *partition(EDGE *left, EDGE *right, float pivot)
+EDGE *partition(EDGE *left, EDGE *right, double pivot)
 {
   while (left <= right)
     {
@@ -70,7 +70,7 @@ EDGE *partition(EDGE *left, EDGE *right, float pivot)
 void quicker_sort(EDGE *left, EDGE *right)
 {
   EDGE *p;
-  float pivot;
+  double pivot;
 
   if (find_pivot(left, right, &pivot) == yes)
     {
@@ -84,10 +84,10 @@ void quicker_sort(EDGE *left, EDGE *right)
 //--------------------start initialize pixels ----------------------------------
 //initialize pixels. See the explination of the pixel class above.
 //initially every pixel is assumed to belong to a group consisting of only itself
-void  initialisePIXELs(float *wrapped_image, unsigned char *input_mask, unsigned char *extended_mask, PIXELM *pixel, int image_width, int image_height)
+void  initialisePIXELs(double *wrapped_image, unsigned char *input_mask, unsigned char *extended_mask, PIXELM *pixel, int image_width, int image_height)
 {
   PIXELM *pixel_pointer = pixel;
-  float *wrapped_image_pointer = wrapped_image;
+  double *wrapped_image_pointer = wrapped_image;
   unsigned char *input_mask_pointer = input_mask;
   unsigned char *extended_mask_pointer = extended_mask;
   int i, j;
@@ -117,9 +117,9 @@ void  initialisePIXELs(float *wrapped_image, unsigned char *input_mask, unsigned
 //-------------------end initialize pixels -----------
 
 //gamma function in the paper
-float wrap(float pixel_value)
+double wrap(double pixel_value)
 {
-  float wrapped_pixel_value;
+  double wrapped_pixel_value;
   if (pixel_value > PI)	wrapped_pixel_value = pixel_value - TWOPI;
   else if (pixel_value < -PI) wrapped_pixel_value = pixel_value + TWOPI;
   else wrapped_pixel_value = pixel_value;
@@ -127,9 +127,9 @@ float wrap(float pixel_value)
 }
 
 // pixelL_value is the left pixel,	pixelR_value is the right pixel
-int find_wrap(float pixelL_value, float pixelR_value)
+int find_wrap(double pixelL_value, double pixelR_value)
 {
-  float difference;
+  double difference;
   int wrap_value;
   difference = pixelL_value - pixelR_value;
 
@@ -252,15 +252,15 @@ void extend_mask(unsigned char *input_mask, unsigned char *extended_mask,
     }
 }
 
-void calculate_reliability(float *wrappedImage, PIXELM *pixel,
+void calculate_reliability(double *wrappedImage, PIXELM *pixel,
 			   int image_width, int image_height,
 			   params_t *params)
 {
   int image_width_plus_one = image_width + 1;
   int image_width_minus_one = image_width - 1;
   PIXELM *pixel_pointer = pixel + image_width_plus_one;
-  float *WIP = wrappedImage + image_width_plus_one; //WIP is the wrapped image pointer
-  float H, V, D1, D2;
+  double *WIP = wrappedImage + image_width_plus_one; //WIP is the wrapped image pointer
+  double H, V, D1, D2;
   int i, j;
 
   for (i = 1; i < image_height -1; ++i)
@@ -286,7 +286,7 @@ void calculate_reliability(float *wrappedImage, PIXELM *pixel,
     {
       //calculating the reliability for the left border of the image
       PIXELM *pixel_pointer = pixel + image_width;
-      float *WIP = wrappedImage + image_width;
+      double *WIP = wrappedImage + image_width;
 
       for (i = 1; i < image_height - 1; ++i)
 	{
@@ -325,7 +325,7 @@ void calculate_reliability(float *wrappedImage, PIXELM *pixel,
     {
       //calculating the reliability for the top border of the image
       PIXELM *pixel_pointer = pixel + 1;
-      float *WIP = wrappedImage + 1;
+      double *WIP = wrappedImage + 1;
 
       for (i = 1; i < image_width - 1; ++i)
 	{
@@ -571,7 +571,7 @@ void  unwrapImage(PIXELM *pixel, int image_width, int image_height)
 
   for (i = 0; i < image_size; i++)
     {
-      pixel_pointer->value += TWOPI * (float)(pixel_pointer->increment);
+      pixel_pointer->value += TWOPI * (double)(pixel_pointer->increment);
       pixel_pointer++;
     }
 }
@@ -581,7 +581,7 @@ void  maskImage(PIXELM *pixel, unsigned char *input_mask, int image_width, int i
 {
   PIXELM *pointer_pixel = pixel;
   unsigned char *IMP = input_mask;	//input mask pointer
-  float min=99999999.f;
+  double min=99999999.f;
   int i;
   int image_size = image_width * image_height;
 
@@ -615,11 +615,11 @@ void  maskImage(PIXELM *pixel, unsigned char *input_mask, int image_width, int i
 //phase map.  copy the image on the buffer passed to this unwrapper to
 //over-write the unwrapped phase map on the buffer of the wrapped
 //phase map.
-void  returnImage(PIXELM *pixel, float *unwrapped_image, int image_width, int image_height)
+void  returnImage(PIXELM *pixel, double *unwrapped_image, int image_width, int image_height)
 {
   int i;
   int image_size = image_width * image_height;
-  float *unwrapped_image_pointer = unwrapped_image;
+  double *unwrapped_image_pointer = unwrapped_image;
   PIXELM *pixel_pointer = pixel;
 
   for (i=0; i < image_size; i++)
@@ -631,7 +631,7 @@ void  returnImage(PIXELM *pixel, float *unwrapped_image, int image_width, int im
 }
 
 //the main function of the unwrapper
-void unwrap2D(float* wrapped_image, float* UnwrappedImage, unsigned char* input_mask,
+void unwrap2D(double *wrapped_image, double *UnwrappedImage, unsigned char *input_mask,
 	 int image_width, int image_height,
 	 int wrap_around_x, int wrap_around_y)
 {
