+++ BrillouinAcquisition/external/unwrap/unwrap2D.h
@@ -7,7 +7,7 @@
 //published in the Journal Applied Optics, Vol. 41, No. 35, pp. 7437, 2002.
 //This program was written by Munther Gdeisat, Liverpool John Moores University, United Kingdom.
 //Date 26th August 2007
-//The wrapped phase map is assumed to be of floating point data type. The resultant unwrapped phase map is also of floating point type.
+//The wrapped phase map is assumed to be of double data type. The resultant unwrapped phase map is also of double type.
 //The mask is of byte data type.
 //When the mask is 1 this means that the pixel is valid
 //When the mask is 0 this means that the pixel is invalid (noisy or corrupted pixel)
@@ -23,15 +23,15 @@
 //TODO: remove global variables
 //TODO: make thresholds independent
 
-static float PI = 3.141592654f;
-static float TWOPI = 6.283185307f;
+static double PI = 3.141592654f;
+static double TWOPI = 6.283185307f;
 
 #define NOMASK 0
 #define MASK 1
 
 typedef struct
 {
-  float mod;
+  double mod;
   int x_connectivity;
   int y_connectivity;
   int no_of_edges;
@@ -42,8 +42,8 @@ struct PIXELM
 {
   int increment;		//No. of 2*pi to add to the pixel to unwrap it
   int number_of_pixels_in_group;//No. of pixel in the pixel group
-  float value;			//value of the pixel
-  float reliability;
+  double value;			//value of the pixel
+  double reliability;
   unsigned char input_mask;	//0 pixel is masked. NOMASK pixel is not masked
   unsigned char extended_mask;	//0 pixel is masked. NOMASK pixel is not masked
   int group;			//group No.
@@ -59,7 +59,7 @@ typedef struct PIXELM PIXELM;
 //if we have S pixels, then we have S horizontal edges and S vertical edges
 struct EDGE
 {
-  float reliab;			//reliabilty of the edge and it depends on the two pixels
+  double reliab;			//reliabilty of the edge and it depends on the two pixels
   PIXELM *pointer_1;		//pointer to the first pixel
   PIXELM *pointer_2;		//pointer to the second pixel
   int increment;		//No. of 2*pi to add to one of the pixels to
@@ -76,24 +76,24 @@ typedef struct EDGE EDGE;
 
 typedef enum {yes, no} yes_no;
 
-yes_no find_pivot(EDGE *left, EDGE *right, float *pivot_ptr);
+yes_no find_pivot(EDGE *left, EDGE *right, double *pivot_ptr);
 
-EDGE *partition(EDGE *left, EDGE *right, float pivot);
+EDGE *partition(EDGE *left, EDGE *right, double pivot);
 
 void quicker_sort(EDGE *left, EDGE *right);
 
-void initialisePIXELs(float *wrapped_image, unsigned char *input_mask, unsigned char *extended_mask, PIXELM *pixel, int image_width, int image_height);
+void initialisePIXELs(double *wrapped_image, unsigned char *input_mask, unsigned char *extended_mask, PIXELM *pixel, int image_width, int image_height);
 
-float wrap(float pixel_value);
+double wrap(double pixel_value);
 
-int find_wrap(float pixelL_value, float pixelR_value);
+int find_wrap(double pixelL_value, double pixelR_value);
 
 void extend_mask(unsigned char *input_mask, unsigned char *extended_mask,
 	int image_width, int image_height,
 	params_t *params);
 
 
-void calculate_reliability(float *wrappedImage, PIXELM *pixel,
+void calculate_reliability(double *wrappedImage, PIXELM *pixel,
 	int image_width, int image_height,
 	params_t *params);
 
@@ -111,9 +111,9 @@ void unwrapImage(PIXELM *pixel, int image_width, int image_height);
 
 void maskImage(PIXELM *pixel, unsigned char *input_mask, int image_width, int image_height);
 
-void returnImage(PIXELM *pixel, float *unwrapped_image, int image_width, int image_height);
+void returnImage(PIXELM *pixel, double *unwrapped_image, int image_width, int image_height);
 
-void unwrap2D(float* wrapped_image, float* UnwrappedImage, unsigned char* input_mask,
+void unwrap2D(double *wrapped_image, double *UnwrappedImage, unsigned char *input_mask,
 	int image_width, int image_height,
 	int wrap_around_x, int wrap_around_y);
